refactor(core): share graph hint conversion
